backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-docx==1.1.0
python-multipart==0.0.9
google-generativeai==0.3.2
aiofiles==23.2.1